requirements.txt
mediapipe
opencv-python>=4.8.0
pyautogui>=0.9.54
numpy>=1.24.0
pystray>=0.19.0
Pillow>=10.0.0
